+++ src/c/profile/transport/tcp/tcp_transport_windows.c
@@ -46,21 +46,21 @@ bool uxr_close_tcp_platform(struct uxrTCPPlatform* platform)
     return rv;
 }
 
-uint16_t uxr_write_tcp_data_platform(struct uxrTCPPlatform* platform, const uint8_t* buf, uint16_t len)
+size_t uxr_write_tcp_data_platform(struct uxrTCPPlatform* platform, const uint8_t* buf, size_t len)
 {
-    uint16_t rv = 0;
+    size_t rv = 0;
     int bytes_sent = send(platform->poll_fd.fd, (const char*)buf, (int)len, 0);
     if (0 < bytes_sent)
     {
         // TODO (julian): take into account errors.
-        rv = (uint16_t)bytes_sent;
+        rv = (size_t)bytes_sent;
     }
     return rv;
 }
 
-uint16_t uxr_read_tcp_data_platform(struct uxrTCPPlatform* platform, uint8_t* buf, uint16_t len, int timeout)
+size_t uxr_read_tcp_data_platform(struct uxrTCPPlatform* platform, uint8_t* buf, size_t len, int timeout)
 {
-    uint16_t rv = 0;
+    size_t rv = 0;
     int poll_rv = WSAPoll(&platform->poll_fd, 1, timeout);
     if (0 < poll_rv)
     {
@@ -68,7 +68,7 @@ uint16_t uxr_read_tcp_data_platform(struct uxrTCPPlatform* platform, uint8_t* bu
         if (0 < bytes_received)
         {
             // TODO (julian): take into account errors.
-            rv = (uint16_t)bytes_received;
+            rv = (size_t)bytes_received;
         }
     }
     return rv;
